B: keep streaming logs after process is aborted
I was passing in ctx to container.Run/container.Attach, which I learned
results in the logs after that point to not be streamed. so, when e.g. a
task is aborted, it'll be sent a SIGTERM, but if we cancel the ctx used
in container.Run, we won't see the logs of that process after that
point.

I decided to still pass in a ctx to runtime.Container.Run/Attach since
other implementations may be able to make use of the ctx more easily,
but the intention of that context is solely for the initial request. we
can evaluate whether or not passing in a context makes sense in the
future.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>