add upstream causality
before, it would only figure out the tree of resources and builds that
were generated as a result of the root resource version, now it also figures
out the builds and resources that resulted in the creation of the root
resource version.

the api response now returns 2 separate trees joined at the root node,
the root.outputOf represents the upstream and root.inputTo is the
downstream.

Signed-off-by: Bohan Chen <bochen@pivotal.io>