containerd: add hostname to appropriate files
The hostname is the container's handle. This is now added to /etc/hostname.

Both the hostname and the container's IP is appended to /etc/hosts. The
container's IP is only available once the network has been initialized.
Therefore, the original /etc/hosts has to be updated after. The /etc/hosts on
the container is bindmounted to a file on the worker.  We simply update this
file directly.

issue#6811

Co-authored-by: Muntasir Chowdhury <mchowdhury@pivotal.io>
Signed-off-by: Bohan Chen <bochen@pivotal.io>